libsel4vm: Add multikernel callback mechanism
Note that this commit should still be treated as a draft.

This supports a multiserver vmm architecture where an SMP guest is
supported via isolated single-core vmm applications each managing a
single core and coordinating to support the single-system-image
abstraction.

This new callback added adds a way for the vmm code to send one-way
messages out to be forwarded to remote cores with
inter-processor-interrupts.

It is expected that the runtime hosting this libsel4vm library is able
to implement the handlers for these IPI messages and make calls into the
library to perform the required updates.

There's only 3 operations supported currently:
- SGI to a remote core
- PSCI power on a core and set its start address,
- Set an SPI target core value for the interrupt controller.