sel4vm,zynqmp: Add multikernel support to vcons
Update the zynqmp's virtual uart to work in a multikernel build.

This sets up a shared mapping of virtual registers and shared device
private state. There's only a single tx and rx fifo stream across all 4
cores, so implement it with shared memory. Use a spinlock for the
synchronization. This could be slow if the vpcu thread is preempted
while holding the lock, but its expected that vmm cores are all
scheduled together using a domain scheduler.

RX and TX to actual hardware have to go through the VM resident on
physical core 0 at the moment and just uses an unmodified serial server
interface.